Cleaned up some things and added an IOverride SetProperties and a RemoveChildren. Both of which work for multiple settings. RemoveChildren will take an array of children to remove and SetProperties accepts an array for both names and values. This eliminates the need to have multiple items when a state is working off the same object.
EX:: <SetProperties target="{crc002}" names="[width,height]" values="[75%,75%]"/>

Also updated  SetProperties and SetProperty to add support for % based values for width and height. In which case percentWidth and percentHeight are updated instead of width and height if a percent type value is specified. Setting these values as a percent will not take advantage of the PercentProxy setting so this change was required to support that through states.